+++ cdn/p5.css
@@ -9,8 +9,12 @@ html {
 }
 
 body {
-  background-color: #dba159;
+  /* background-color: #dba159; */
   height: 100vh;
+  background-color: #330055;
+  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 100 1000'%3E%3Cg %3E%3Ccircle fill='%23330055' cx='50' cy='0' r='50'/%3E%3Cg fill='%233a015d' %3E%3Ccircle cx='0' cy='50' r='50'/%3E%3Ccircle cx='100' cy='50' r='50'/%3E%3C/g%3E%3Ccircle fill='%23410165' cx='50' cy='100' r='50'/%3E%3Cg fill='%2348026e' %3E%3Ccircle cx='0' cy='150' r='50'/%3E%3Ccircle cx='100' cy='150' r='50'/%3E%3C/g%3E%3Ccircle fill='%23500376' cx='50' cy='200' r='50'/%3E%3Cg fill='%2357047e' %3E%3Ccircle cx='0' cy='250' r='50'/%3E%3Ccircle cx='100' cy='250' r='50'/%3E%3C/g%3E%3Ccircle fill='%235f0587' cx='50' cy='300' r='50'/%3E%3Cg fill='%2367068f' %3E%3Ccircle cx='0' cy='350' r='50'/%3E%3Ccircle cx='100' cy='350' r='50'/%3E%3C/g%3E%3Ccircle fill='%236f0798' cx='50' cy='400' r='50'/%3E%3Cg fill='%237707a0' %3E%3Ccircle cx='0' cy='450' r='50'/%3E%3Ccircle cx='100' cy='450' r='50'/%3E%3C/g%3E%3Ccircle fill='%238008a9' cx='50' cy='500' r='50'/%3E%3Cg fill='%238909b1' %3E%3Ccircle cx='0' cy='550' r='50'/%3E%3Ccircle cx='100' cy='550' r='50'/%3E%3C/g%3E%3Ccircle fill='%239109ba' cx='50' cy='600' r='50'/%3E%3Cg fill='%239a09c3' %3E%3Ccircle cx='0' cy='650' r='50'/%3E%3Ccircle cx='100' cy='650' r='50'/%3E%3C/g%3E%3Ccircle fill='%23a309cb' cx='50' cy='700' r='50'/%3E%3Cg fill='%23ad09d4' %3E%3Ccircle cx='0' cy='750' r='50'/%3E%3Ccircle cx='100' cy='750' r='50'/%3E%3C/g%3E%3Ccircle fill='%23b608dc' cx='50' cy='800' r='50'/%3E%3Cg fill='%23c007e5' %3E%3Ccircle cx='0' cy='850' r='50'/%3E%3Ccircle cx='100' cy='850' r='50'/%3E%3C/g%3E%3Ccircle fill='%23c905ee' cx='50' cy='900' r='50'/%3E%3Cg fill='%23d303f6' %3E%3Ccircle cx='0' cy='950' r='50'/%3E%3Ccircle cx='100' cy='950' r='50'/%3E%3C/g%3E%3Ccircle fill='%23D0F' cx='50' cy='1000' r='50'/%3E%3C/g%3E%3C/svg%3E");
+  background-attachment: fixed;
+  background-size: contain;
 }
 
 #render {
@@ -25,7 +29,7 @@ body {
 }
 
 ::-webkit-scrollbar-thumb {
-  background: #dba159;
+  background: #c347ff85;
   border-radius: 20px;
 }
 
@@ -42,15 +46,19 @@ body {
   max-height: 566px;
   width: 100%;
   height: 100%;
-  background-color: #efd780;
-  box-shadow: 0px 0px 3.2px rgb(0 0 0 / 2%), 0px 0px 7.7px rgb(0 0 0 / 3%),
-    0px 0px 14.5px rgb(0 0 0 / 4%), 0px 0px 25.9px rgb(0 0 0 / 4%),
-    0px 0px 48.5px rgb(0 0 0 / 5%), 0px 0px 116px rgb(0 0 0 / 7%);
-  border-radius: 33px;
   text-align: center;
   display: flex;
   align-items: center;
   flex-direction: column;
+
+  background: rgba(53, 53, 53, 0.23);
+  border: 1px solid rgb(186 0 255 / 17%);
+  box-sizing: border-box;
+  box-shadow: inset 0px 0px 9px 3px rgb(0 0 0 / 5%);
+  backdrop-filter: blur(7px);
+  border-radius: 33px;
+
+  border-radius: 33px;
 }
 
 .messages__container {
@@ -60,10 +68,9 @@ body {
   max-height: 264px;
   width: 100%;
   height: 100%;
-
-  background: #efd780;
-  border: 3px solid #9f8f53;
-  box-shadow: 0px 4px 47px 5px rgba(0, 0, 0, 0.25);
+  background: #441657c2;
+  /* border: 3px solid #9f8f53; */
+  box-shadow: 0px 4px 47px 5px rgb(0 0 0 / 25%);
   border-radius: 10px;
   padding: 20px;
   margin-top: 60px;
@@ -77,11 +84,7 @@ body {
   font-weight: bold;
   font-size: 18px;
   line-height: 21px;
-  /* identical to box height */
-
-  /* Dark semi-brown */
-
-  color: #9f8f53;
+  color: #db3fff;
 }
 
 .message__content {
@@ -90,7 +93,7 @@ body {
   font-weight: bold;
   font-size: 14px;
   line-height: 16px;
-  color: #7d7140;
+  color: #c753ff;
   max-width: 80%;
   word-break: break-word;
   text-align: left;
@@ -103,14 +106,14 @@ body {
 }
 
 .title {
+  margin: 0;
   font-style: normal;
   font-weight: bold;
   font-size: 54px;
   line-height: 63px;
-
-  color: #9f8f53;
-  text-shadow: 2px 0 #bbaa6c;
-  margin-bottom: 70px;
+  background: -webkit-linear-gradient(1turn, #cb5eff, #b720ff);
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
 }
 
 .inputs {
@@ -120,6 +123,7 @@ body {
   height: 100%;
   width: 100%;
   row-gap: 10px;
+  justify-content: center;
 }
 
 .input {
@@ -128,9 +132,6 @@ body {
   height: 100%;
   width: 100%;
 
-  background-color: #9f8f53ad;
-
-  border: 1px solid #fcfdaf;
   box-sizing: border-box;
   border-radius: 11px;
   padding: 0 24px;
@@ -138,55 +139,58 @@ body {
   font-style: normal;
   font-weight: 500;
   font-size: 18px;
-  color: rgba(254, 255, 173, 0.973);
+  color: #e266ff;
+
+  border: 0;
+  background-color: rgba(31, 31, 31, 0.37);
+  border-radius: 11px;
   outline: 0;
-  transition: box-shadow 0.3s cubic-bezier(0.455, 0.03, 0.515, 0.955);
+  transition: background-color 0.3s cubic-bezier(0.455, 0.03, 0.515, 0.955);
 }
 
 .input:focus,
 .input:hover {
-  box-shadow: inset 0 0 10px #3f3f3f47;
+  background-color: rgb(31 31 31 / 28%);
 }
 
 .input::placeholder {
   font-style: normal;
   font-weight: 500;
   font-size: 18px;
-  color: rgba(254, 255, 173, 0.72);
+  color: #e266ff;
 }
 
+/* Cool Button */
 .submit {
   max-width: 419px;
   max-height: 63px;
   width: 100%;
   height: 100%;
-  background-color: #55945b;
-  border: 1px solid #3e7944;
-  box-sizing: border-box;
   border-radius: 11px;
-
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: 500;
   font-size: 24px;
   line-height: 28px;
-
-  color: #7bdd85;
+  border: 0;
+  color: #6a0b81;
+  font-weight: 600;
   cursor: pointer;
   margin-top: 20px;
-  box-shadow: inset 0 0 10px #3f3f3f47, 0px 5px 20px 0px #3f6d4462,
-    0px 5px 41px 7px #3f6d443f;
-  transition: box-shadow 0.3s ease-in, background-color 0.3s ease-in,
-    color 0.3s ease-in;
+  box-shadow: inset 0 0 10px #3f3f3f47, 0px 5px 20px 0px #9109ba87,
+    0px 5px 41px 7px #9109ba73;
+
+  background-color: #ad09d4;
+  background-image: linear-gradient(1turn, #ad09d4, #9e07c2);
+  transition: box-shadow 0.3s ease-in, background-image 0.3s ease-in,
+    color 0.3s ease-in, background-color 0.3s ease-in;
 }
 
 .submit:hover {
-  transition: box-shadow 0.17s ease-out, background-color 0.17s ease-out,
-    color 0.17s ease-out;
-  box-shadow: inset 0 0 10px #e6e6e647, 0px 5px 20px 0px #3f6d4462,
-    0px 5px 41px 7px #3f6d443f;
-  background-color: #55945bad;
-  color: #307136;
+  transition: box-shadow 0.17s ease-out, background-image 0.17s ease-out,
+    background-color 0.17s ease-out, color 0.17s ease-out;
+  box-shadow: inset 0 0 10px #fbfbfb14, 0px 5px 20px 0px #900fb2,
+    0px 5px 41px 7px #9109ba;
+  background-color: #ad09d454;
+  background-image: linear-gradient(#ab09d456, #ad09d454);
+  color: #d830ff;
 }
 
 .hint {
@@ -195,7 +199,7 @@ body {
   font-size: 18px;
   line-height: 21px;
   margin-top: 0.9em;
-  color: #9f8f53;
+  color: #de48ff;
 }
 
 .hint span {
@@ -204,12 +208,21 @@ body {
   font-size: 18px;
   line-height: 21px;
 
-  color: #55945b;
+  color: #e266ff;
   margin-left: 4px;
   cursor: pointer;
   white-space: nowrap;
+  display: inline-flex;
+  flex-direction: column;
+  row-gap: 3px;
 }
 
-.hint span:hover {
-  color: #3f6d44;
+.line {
+  background-color: #6100734a;
+  width: 100%;
+  height: 4px;
+}
+
+.hint span:hover .line {
+  background-color: #62007375;
 }
